# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MeridianIQ platform API — data spine, compliance rails and consent.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parties/validate-tin:
    post:
      operationId: validateTin
      tags: [parties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TinCheckInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierCheck"
  /parties/validate-cac:
    post:
      operationId: validateCac
      tags: [parties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CacCheckInput"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IdentifierCheck"
...
  /public/theme:
    get:
      operationId: getPublicTheme
      tags: [whitelabel]
      summary: Resolve firm branding by subdomain (no auth; public shell needs it before login)
      parameters:
        - name: subdomain
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Firm branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FirmTheme"
        "404":
          $ref: "#/components/responses/NotFound"
  /firms/{id}/theme:
    put:
      operationId: updateFirmTheme
      tags: [whitelabel]
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FirmThemeInput"
      responses:
        "200":
          description: Updated firm
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Firm"
        "404":
          $ref: "#/components/responses/NotFound"
  /clients/import:
    post:
      operationId: importClients
      tags: [whitelabel]
      summary: Bulk client import from a practice-management export (CON-05)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientImportInput"
      responses:
        "200":
          description: Per-row import results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientImportResult"
        "400":
          $ref: "#/components/responses/BadRequest"

  # ---- R2: certification portal (CON-05) ----
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Firm:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: string
        name:
          type: string
        subdomain:
          type: ["string", "null"]
        clerkOrgId:
          type: ["string", "null"]
        partyId:
          type: ["string", "null"]
        theme:
          type: ["object", "null"]
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    TinCheckInput:
      type: object
      required: [tin]
      properties:
        tin:
          type: string
    CacCheckInput:
      type: object
      required: [cac]
      properties:
        cac:
          type: string
    IdentifierCheck:
      type: object
      required: [valid, normalized]
      properties:
        valid:
          type: boolean
        normalized:
          type: string
...
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
...
    FirmTheme:
      type: object
      required: [firmId, name, subdomain, theme]
      properties:
        firmId:
          type: string
        name:
          type: string
        subdomain:
          type: ["string", "null"]
        theme:
          type: ["object", "null"]
          additionalProperties: true
    FirmThemeInput:
      type: object
      required: [theme]
      properties:
        subdomain:
          type: string
          pattern: "^[a-z0-9][a-z0-9-]{1,61}[a-z0-9]$"
        theme:
          type: object
          additionalProperties: true
    ClientImportRow:
      type: object
      required: [legalName]
      properties:
        legalName:
          type: string
          minLength: 1
        tin:
          type: string
        cacNumber:
          type: string
        email:
          type: string
        street:
          type: string
        city:
          type: string
        engagementTitle:
          type: string
    ClientImportInput:
      type: object
      required: [rows, commit]
      properties:
        rows:
          type: array
          minItems: 1
          maxItems: 1000
          items:
            $ref: "#/components/schemas/ClientImportRow"
        commit:
          type: boolean
    ClientImportRowResult:
      type: object
      required: [rowNumber, status]
      properties:
        rowNumber:
          type: integer
        status:
          type: string
          enum: [created, exists, invalid]
        partyId:
          type: ["string", "null"]
        engagementId:
          type: ["string", "null"]
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
    ClientImportResult:
      type: object
      required: [rowCount, createdCount, existsCount, invalidCount, committed, rows]
      properties:
        rowCount:
          type: integer
        createdCount:
          type: integer
        existsCount:
          type: integer
        invalidCount:
          type: integer
        committed:
          type: boolean
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ClientImportRowResult"

    # ---- R2 schemas: certification portal (CON-05) ----
...