# lib/api-zod/src/generated/types/clientImportRow.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * MeridianIQ platform API — data spine, compliance rails and consent.
 * OpenAPI spec version: 0.1.0
 */

export interface ClientImportRow {
  /** @minLength 1 */
  legalName: string;
  tin?: string;
  cacNumber?: string;
  email?: string;
  street?: string;
  city?: string;
  engagementTitle?: string;
}